+++ packages/grafana-prometheus/src/components/metrics-browser/LabelSelector.tsx
@@ -4,7 +4,7 @@ import { selectors } from '@grafana/e2e-selectors';
 import { Trans, t } from '@grafana/i18n';
 import { BrowserLabel as PromLabel, Input, Label, useStyles2 } from '@grafana/ui';
 
-import { METRIC_LABEL } from '../../types';
+import { METRIC_LABEL } from '../../constants';
 
 import { useMetricsBrowser } from './MetricsBrowserContext';
 import { getStylesLabelSelector } from './styles';
