+++ packages/grafana-prometheus/src/components/metrics-browser/MetricSelector.tsx
@@ -5,8 +5,9 @@ import { selectors } from '@grafana/e2e-selectors';
 import { Trans, t } from '@grafana/i18n';
 import { BrowserLabel as PromLabel, Input, Label, useStyles2 } from '@grafana/ui';
 
+import { LIST_ITEM_SIZE } from '../../constants';
+
 import { useMetricsBrowser } from './MetricsBrowserContext';
-import { LIST_ITEM_SIZE } from './constants';
 import { getStylesMetricSelector } from './styles';
 
 export function MetricSelector() {
