--- packages/grafana-prometheus/src/components/metrics-browser/constants.ts
@@ -1,2 +1,0 @@
-export const LIST_ITEM_SIZE = 25;
-export const LAST_USED_LABELS_KEY = 'grafana.datasources.prometheus.browser.labels';
