+++ packages/grafana-prometheus/src/components/metrics-browser/useMetricsLabelsValues.test.ts
@@ -2,11 +2,10 @@ import { act, renderHook, waitFor } from '@testing-library/react';
 
 import { TimeRange } from '@grafana/data';
 
+import { DEFAULT_SERIES_LIMIT, EMPTY_SELECTOR, LAST_USED_LABELS_KEY, METRIC_LABEL } from '../../constants';
 import { PrometheusLanguageProviderInterface } from '../../language_provider';
 import { getMockTimeRange } from '../../test/__mocks__/datasource';
-import { DEFAULT_SERIES_LIMIT, EMPTY_SELECTOR, METRIC_LABEL } from '../../types';
 
-import { LAST_USED_LABELS_KEY } from './constants';
 import * as selectorBuilderModule from './selectorBuilder';
 import { useMetricsLabelsValues } from './useMetricsLabelsValues';
