+++ packages/grafana-prometheus/src/types.ts
@@ -189,10 +189,3 @@ export type RecordingRuleIdentifier = {
   identifier?: string;
   identifierValue?: string;
 };
-
-export const EMPTY_SELECTOR = '{}';
-export const DEFAULT_SERIES_LIMIT = '40000';
-export const REMOVE_SERIES_LIMIT = 'none';
-export const MATCH_ALL_LABELS_STR = '__name__!=""';
-export const MATCH_ALL_LABELS = '{__name__!=""}';
-export const METRIC_LABEL = '__name__';
